requirements.txt
qrcode[pil]>=7.4.2
Pillow>=10.0.0
requests>=2.31.0
click>=8.1.0
pytest>=7.4.0
pytest-cov>=4.1.0
boto3>=1.34.0
webdav4>=0.9.8
psutil>=5.9.0